Update Count Commas in Range II solution

+++ 4248-count-commas-in-range-ii/4248-count-commas-in-range-ii.cpp
@@ -2,9 +2,9 @@ class Solution {
 public:
     long long countCommas(long long n) {
         long long cnt = 0;
-        if( (long long)n >= (long long)1000000000000000) {
-            cnt += (long long)((long long)5*(long long)((long long)n-(long long)999999999999999));
-            n = (long long)999999999999999;
+        if(n >= 1000000000000000) {
+            cnt += 5*(n-999999999999999);
+            n = 999999999999999;
         }
         if( (long long)n >= (long long)1000000000000) {
             cnt += (long long)((long long)4*(long long)((long long)n-(long long)999999999999));
